updated blackjack project
inspired by the model code they gave. 1. Didnt use the blackjack = 0 because I determined that even if the score == 21 and len(hand) > 2, the game_active flag would still be set as false so it wasn't necessary to specify ending the game if you draw blackjack from the start as it would be checked to the end.
2. I changed the rules of the game such that even if computer draws blackjack from the start you wont automatically lose even if you blackjack, you would draw ( official rules, not sure why this project made it such that computer blackjack means you lose even if u get blackjack )